refactor(라우팅): router의 render부분을 views로 분리
