+++ src/js/views/index.js
@@ -5,9 +5,13 @@ import { renderLines } from './contents/lines/index.js';
 import { renderSections } from './contents/sections/index.js';
 import { renderLogin } from './auths/login/index.js';
 import { renderSignUp } from './auths/signUp/index.js';
-import { ROUTES } from '../constants/index.js';
+import { $ } from '../utils/index.js';
+import { ROUTES, TITLES } from '../constants/index.js';
 
-const render = {
+const $header = $('header');
+const $main = $('main');
+
+const renderContents = {
   [ROUTES.HOME]: renderHome,
   [ROUTES.STATIONS]: renderStations,
   [ROUTES.LINES]: renderLines,
@@ -17,4 +21,16 @@ const render = {
   [ROUTES.LOGOUT]: renderHome,
 };
 
-export { renderHeader, renderHome, render };
+const renderTitle = (pathname) => {
+  document.title = TITLES[pathname];
+};
+
+export const renderContent = (pathname) => {
+  renderContents[pathname]($main);
+};
+
+export const render = (pathname) => {
+  renderTitle(pathname);
+  renderHeader($header);
+  renderContent(pathname);
+};
